Plugin rewrite
Some context first:

When I implemented the first version of the plugin years ago, I decided
to use Python's default HTML parser in order not to force users to pull
down any external parser.

It was definitely a good learning experience but the default parser just
plainly sucks since there is no concept of a tree, which you either have
to build yourself or do without. The latter is what I did since the code
was simple enough at first.

It grew more complicated over the years as I fixed bugs and made it
evolve following RuTracker itself evolving, and became a huge pain to
maintain.

This plugin was also one of my first forays into integrating with an
external program via API, as a young programmer. Again, good learning
experience, but I also missed obvious stuff.

One such instance was the fix in 24fe4911a5a6000f4564babd22248750be04d6cf:
when initially implementing, I was under the impression results should
be consolidated and printed all at once at the end, akin to a function
return value. This made the plugin quite slow because all pages of
results had to be parsed before being printed at once.

Now that I'm looking over the API again, it feels obvious that it was
designed in a way that makes it possible for asynchronous retrievals to
return results as they are coming in. Obviously, this is perfect for
parallelization, so I happily decided to hack in a `ThreadPoolExecutor`
map so that searches with multiple pages would concurrently retrieve
pages once the first one got through.

Doing that yielded some issues with the monolithic parser, which is when
I quickly decided I didn't want to deal with that anymore. Hence the
rewrite.

This is the first pass, I'll do some more passes to refactor other old
ugly pieces, such as the try/except blocks for open() methods. Already
works great as-is though :)

- Add concurrent fetches for searches with multiple pages.
- Replace complicated HTML parser with simple regular expression.
- Use timeit for timing tests.